fix SN install
git-svn-id: https://svn.code.sf.net/p/xcat/code/xcat-core/trunk@2520 8638fb3e-16cb-4fca-ae20-7b5d299a9bcd

+++ xCAT/xCAT.spec
@@ -122,8 +122,9 @@ fi
 chkconfig vsftpd on
 /etc/init.d/vsftpd start
 # remove any service node file
+if [ -f /etc/xCATSN ]; then
  rm  /etc/xCATSN
-
+fi
 if [ "$1" = "1" ]; then #Only if installing for the first time..
     mkdir -p /root/.ssh
     chmod 700 /root/.ssh
